Move partitioning logic to users
